+++ src/muibook/story/parts/visual.js
@@ -16,18 +16,25 @@ class partsVisual extends HTMLElement {
         >
 
           <spec-card title="Background">
-            <story-part-slat slot="body" token="part(background)" usage="Link, Button, Select" output="Defines the background of the element." ></story-part-slat>
+            <story-part-slat slot="body" token="part(background)" usage="Link, Button, Select, Drawer" output="Defines the background of the element." ></story-part-slat>
             <mui-code slot="footer">
               mui-link::part(background) {
               <br />
               &nbsp;&nbsp;background: var(--white-opacity-0);
               <br />
               }
+              <br />
+              <br />
+              mui-drawer::part(background) {
+              <br />
+              &nbsp;&nbsp;background: var(--app-navbar-surface);
+              <br />
+              }
             </mui-code>
           </spec-card>
 
           <spec-card title="Border">
-            <story-part-slat slot="body" token="part(border)" usage="Link, Button, Select" output="Defines the border of the element."></story-part-slat>
+            <story-part-slat slot="body" token="part(border)" usage="Link, Button, Select, Drawer" output="Defines the border of the element."></story-part-slat>
             <mui-code slot="footer">
               mui-link::part(border) {
               <br />
@@ -41,7 +48,7 @@ class partsVisual extends HTMLElement {
           </spec-card>
 
           <spec-card title="Border Radius">
-            <story-part-slat slot="body" token="part(border-radius)" usage="Link, Button, Select" output="Defines the border radius of the element."></story-part-slat>
+            <story-part-slat slot="body" token="part(border-radius)" usage="Link, Button, Select, Drawer" output="Defines the border radius of the element."></story-part-slat>
             <mui-code slot="footer">
               mui-link::part(border-radius) {
               <br />
@@ -52,7 +59,7 @@ class partsVisual extends HTMLElement {
           </spec-card>
 
           <spec-card title="Box Shadow">
-            <story-part-slat slot="body" token="part(box-shadow)" usage="Link, Button, Select" output="Defines the box-shadow of the element."></story-part-slat>
+            <story-part-slat slot="body" token="part(box-shadow)" usage="Link, Button, Select, Drawer" output="Defines the box-shadow of the element."></story-part-slat>
             <mui-code slot="footer">
               mui-link::part(box-shadow) {
               <br />
@@ -63,7 +70,7 @@ class partsVisual extends HTMLElement {
           </spec-card>
 
           <spec-card title="Opacity">
-            <story-part-slat slot="body" token="part(opacity)" usage="Link, Button, Select" output="Defines the opacity of the element."></story-part-slat>
+            <story-part-slat slot="body" token="part(opacity)" usage="Link, Button, Select, Drawer" output="Defines the opacity of the element."></story-part-slat>
             <mui-code slot="footer">
               mui-link::part(opacity) {
               <br />
@@ -74,7 +81,7 @@ class partsVisual extends HTMLElement {
           </spec-card>
 
           <spec-card title="Transition">
-            <story-part-slat slot="body" token="part(transition)" usage="Link, Button, Select" output="Defines the transition of the element when animation is used."></story-part-slat>
+            <story-part-slat slot="body" token="part(transition)" usage="Link, Button, Select, Drawer" output="Defines the transition of the element when animation is used."></story-part-slat>
             <mui-code slot="footer"  scrollable>
               mui-link::part(transition) {
               <br />
@@ -85,7 +92,7 @@ class partsVisual extends HTMLElement {
           </spec-card>
 
           <spec-card title="Outline">
-            <story-part-slat slot="body" token="part(outline)" usage="Link, Button, Select" output="Defines the outline of the element, typically for focus states."></story-part-slat>
+            <story-part-slat slot="body" token="part(outline)" usage="Link, Button, Select, Drawer" output="Defines the outline of the element, typically for focus states."></story-part-slat>
             <mui-code slot="footer">
               mui-link::part(outline) {
               <br />